Fix 'DateRange.rangeWithMinimumSize()'
Before, 'DateRange.rangeWithMinimumSize()' possibly returned a
range that was partially outside 'outerRange'. This allowed the
user to select a range that ended later than what was shown in
the calendar component.

Now the function returns 'DateRange.emptyRange()' if it is not
possible to return a range with the specified minimum size
starting at the given date.

This also prevents an infinite loop if '{minimumRange: 7,
disableWeekends: true}' and instant selection of the last week
of the calendar is attempted.