Fix 'createCalendar()'
Before, 'setYearLabel()' was invoked in 'createCalendar()' for
both non-popup and popup calendars. For popup calendars, however,
it is not possible to calculate the current year, for it depends
on the average cell height.

Now 'setYearLabel()' is invoked in 'createCalendar()' only for
non-popup calendars. For popup calendars, it is invoked in
'toggleCalendar()' after calculating the average cell height.

This fixes the regression introduced in 95d4bd.