Add option 'selectWeek'
The option 'selectWeek' triggers a behavior similar to the normal
non-range behavior but clicking a date causes the entire week to
be selected. Because the selection is a range, 'selectWeek'
requires the 'startDate' and 'endDate' input fields. If the
option 'disableWeekends' is true, the weekend is excluded from the
selection.