allow mouse events to start/end on current date -- fixes #35
mousedown, mouseover, mouseup events return the current
date's table cell's inner div, which does not have a mouse
event bound to it, so was throwing an uncaught exception.
this patch include a test and modifies the test utilities to
return the div, so that the test will fail without the patch.